Renamed api folder to utils
*and paths.py to api_paths.py
