+++ resources/lib/kodi/ui/xmldialogs.py
@@ -149,7 +149,7 @@ class ParentalControl(xbmcgui.WindowXMLDialog):
             # # Validate pin length
             # if not self._validate_pin(pin):
             #     return
-            import resources.lib.api.api_requests as api
+            import resources.lib.utils.api_requests as api
             data = {'guid': self.data['profileInfo']['guid'],
                     'experience': self.data['experience'],
                     'maturity': self.rating_levels[self.current_level_index]['value'],
@@ -268,7 +268,7 @@ class RatingThumb(xbmcgui.WindowXMLDialog):
         if controlID in [10010, 10020, 10012, 10022]:  # Rating and close
             rating_map = {10010: 2, 10020: 1, 10012: 0, 10022: 0}
             rating_value = rating_map[controlID]
-            from resources.lib.api.api_requests import rate_thumb
+            from resources.lib.utils.api_requests import rate_thumb
             rate_thumb(self.videoid, rating_value, self.track_id_jaw)
             self.close()
         if controlID in [10040, 100]:  # Close
