+++ resources/lib/services/msl/msl_handler.py
@@ -15,7 +15,7 @@ import time
 import xbmcaddon
 
 import resources.lib.common as common
-from resources.lib.api.exceptions import CacheMiss
+from resources.lib.utils.exceptions import CacheMiss
 from resources.lib.common.cache_utils import CACHE_MANIFESTS
 from resources.lib.database.db_utils import TABLE_SESSION
 from resources.lib.globals import G
