+++ resources/lib/services/nfsession/session/http_requests.py
@@ -12,10 +12,10 @@ from __future__ import absolute_import, division, unicode_literals
 
 import json
 
-import resources.lib.api.website as website
+import resources.lib.utils.website as website
 import resources.lib.common as common
-from resources.lib.api.exceptions import (APIError, WebsiteParsingError, MbrStatusError, MbrStatusAnonymousError,
-                                          HttpError401)
+from resources.lib.utils.exceptions import (APIError, WebsiteParsingError, MbrStatusError, MbrStatusAnonymousError,
+                                            HttpError401)
 from resources.lib.common import cookies
 from resources.lib.database.db_utils import TABLE_SESSION
 from resources.lib.globals import G
